+++ src/rendering/common/common_preamble.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifndef __HLSL_VERSION
+#define _hlsl 0
+#else
+#define _hlsl 1
+#endif
