+++ src/rendering/common/common_structs.cpp
