--- src/shaders/common_structs.hlsli
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "../rendering/common_structs.h"
