+++ src/shaders/global_params.hlsli
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "common_structs.hlsli"
+#include "../rendering/common/common_structs.h"
+#include "../rendering/common/common_registers.h"
 
 cbuffer GlobalParams : REGISTER_B(REGISTER_GLOBAL_PARAMS)
 {
